+++ src/it/resources/transactionTest/chainDump.conf
@@ -1,4 +1,4 @@
-node = "enode://9e56c4eb31089d9b6b40c183f0e549ca5be049dbb22f4b0e25f4955e19d580915ddc32ec26ff38693cbf6954ff743478e42a532b37e3c7a87415067d81c019d5@192.168.1.105:30303"
+node = "enode://9e56c4eb31089d9b6b40c183f0e549ca5be049dbb22f4b0e25f4955e19d580915ddc32ec26ff38693cbf6954ff743478e42a532b37e3c7a87415067d81c019d5@192.168.1.207:30303"
 genesisHash = "c2503fa5d31cc3daf8a97ea57a15d2bee8e867252e2d8d6fdb84eedc17309e55"
 networkId = 146
 startBlock = 0
