+++ src/it/resources/transactionTest/purchaseContract/bodies.txt
