+++ src/it/resources/transactionTest/purchaseContract/contractTrees.txt
