+++ src/it/resources/transactionTest/purchaseContract/evmCode.txt
