+++ src/it/resources/transactionTest/purchaseContract/headers.txt
