+++ src/it/resources/transactionTest/purchaseContract/receipts.txt
