+++ src/it/resources/transactionTest/purchaseContract/stateTree.txt
