+++ src/it/scala/io/iohk/ethereum/txExecTest/util/DumpChainApp.scala
@@ -1,4 +1,4 @@
-package io.iohk.ethereum.transactionTest.util
+package io.iohk.ethereum.txExecTest.util
 
 import akka.actor.ActorSystem
 import akka.agent.Agent
@@ -15,10 +15,7 @@ import org.spongycastle.util.encoders.Hex
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration.FiniteDuration
 
-object DumpChainApp {
-  // scalastyle:off
-  def main(args: Array[String]): Unit = {
-
+object DumpChainApp extends App{
     val conf = ConfigFactory.load("transactionTest/chainDump.conf")
     val node = conf.getString("node")
     val genesisHash = ByteString(Hex.decode(conf.getString("genesisHash")))
@@ -62,7 +59,8 @@ object DumpChainApp {
 
   class BlockchainMock(genesisHash: ByteString) extends Blockchain {
 
-    class FakeHeader() extends BlockHeader(ByteString.empty, ByteString.empty, ByteString.empty, ByteString.empty, ByteString.empty, ByteString.empty, ByteString.empty, 0, 0, 0, 0, 0, ByteString.empty, ByteString.empty, ByteString.empty) {
+    class FakeHeader() extends BlockHeader(ByteString.empty, ByteString.empty, ByteString.empty, ByteString.empty,
+      ByteString.empty, ByteString.empty, ByteString.empty, 0, 0, 0, 0, 0, ByteString.empty, ByteString.empty, ByteString.empty) {
       override lazy val hash: ByteString = genesisHash
     }
 
@@ -93,6 +91,4 @@ object DumpChainApp {
     override def getEvmCodeByHash(hash: ByteString): Option[ByteString] = ???
 
     override def getReceiptsByHash(blockhash: ByteString): Option[Seq[Receipt]] = ???
-  }
-
 }
